move resolveValue function from types.ts to utils.ts
