+++ src/core/utils.ts
@@ -1,3 +1,5 @@
+import { ValueFunction, ValueOrFunction } from "./types";
+
 export const genId = (() => {
   let count = 0;
   return () => {
@@ -28,3 +30,13 @@ export const prefersReducedMotion = (() => {
     return shouldReduceMotion;
   };
 })();
+
+const isFunction = <TValue, TArg>(
+  valOrFunction: ValueOrFunction<TValue, TArg>
+): valOrFunction is ValueFunction<TValue, TArg> =>
+  typeof valOrFunction === 'function';
+
+export const resolveValue = <TValue, TArg>(
+  valOrFunction: ValueOrFunction<TValue, TArg>,
+  arg: TArg
+): TValue => (isFunction(valOrFunction) ? valOrFunction(arg) : valOrFunction);
